--- src/components/CreateModal/index.js
@@ -1,1 +1,0 @@
-export * from './CreateModal';
