+++ src/components/FormDialog/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/export
+export * from './FormDialog';
